split out move and copy
